test: check ready for painting and surface robustly
The X11 window can be ready for painting and having a surface at different
points in time. For that try-verify when the windows are ready for painting
and in case the surface is not yet available spy on the change.

(cherry picked from commit 6691f09e700f465a0ed6562be40b1eb70ec38913)